edit bar to markup, add badge
