+++ packages/server/markup/admin.js
@@ -1,4 +1,13 @@
-const { div, hr, form, input, label, i } = require("@saltcorn/markup/tags");
+const {
+  div,
+  hr,
+  form,
+  input,
+  label,
+  i,
+  a,
+  span,
+} = require("@saltcorn/markup/tags");
 
 const restore_backup = (csrf, inner) =>
   form(
@@ -18,5 +27,25 @@ const restore_backup = (csrf, inner) =>
       onchange: "this.form.submit();",
     })
   );
+const add_edit_bar = ({ role, title, contents, what, url }) => {
+  if (role > 1) return contents;
+  const bar = div(
+    { class: "alert alert-light" },
+    title,
+    what && span({ class: "ml-1 badge badge-primary" }, what),
+    a(
+      {
+        class: "ml-4",
+        href: url,
+      },
+      "Edit&nbsp;",
+      i({ class: "fas fa-edit" })
+    )
+  );
 
-module.exports = { restore_backup };
+  if (contents.above) {
+    contents.above.unshift(bar);
+    return contents;
+  } else return { above: [bar, contents] };
+};
+module.exports = { restore_backup, add_edit_bar };
