+++ packages/server/routes/page.js
@@ -5,29 +5,11 @@ const { div, a, i } = require("@saltcorn/markup/tags");
 const { renderForm } = require("@saltcorn/markup");
 const { getState } = require("@saltcorn/data/db/state");
 const { setTenant, error_catcher } = require("../routes/utils.js");
+const { add_edit_bar } = require("../markup/admin.js");
 
 const router = new Router();
 module.exports = router;
-const add_edit_bar = (role, page, contents) => {
-  if (role > 1) return contents;
-  const bar = div(
-    { class: "alert alert-light" },
-    page.name,
-    a(
-      {
-        class: "ml-4",
-        href: `/pageedit/edit/${encodeURIComponent(page.name)}`,
-      },
-      "Edit&nbsp;",
-      i({ class: "fas fa-edit" })
-    )
-  );
 
-  if (contents.above) {
-    contents.above.unshift(bar);
-    return contents;
-  } else return { above: [bar, contents] };
-};
 router.get(
   "/:pagename",
   setTenant,
@@ -50,7 +32,13 @@ router.get(
         res.sendWrap(
           { title: db_page.title, description: db_page.description } ||
             `${pagename} page`,
-          add_edit_bar(role, db_page, contents)
+          add_edit_bar({
+            role,
+            title: db_page.name,
+            what: req.__("Page"),
+            url: `/pageedit/edit/${encodeURIComponent(db_page.name)}`,
+            contents,
+          })
         );
       } else
         res
